style: More eslint rules
To explain the value of the `generator-star-spacing`:
Usually, `function* name() {}` is preferred as the star is modifying the type of *function* it is.
For anonymous functions, `function* () {}` is preferred as its consistent with the above and doesn't imply multiplication as much as `function*() {}`
For methods, `*name() {}` or `static *name() {}` is preferred as the star has more to do with the method that the fact that its static.